reformatted shared functions
